@@ -10,21 +10,35 @@
 		"lint": "find . -name \\*.php -not -path './vendor/*' -print0 | xargs -0 -n1 php -l",
 		"cs:check": "php-cs-fixer fix --dry-run --diff",
 		"cs:fix": "php-cs-fixer fix",
-		"psalm": "psalm.phar --no-cache",
-		"psalm:update-baseline": "psalm.phar --threads=1 --update-baseline",
-		"psalm:update-baseline:force": "psalm.phar --threads=1 --update-baseline --set-baseline=tests/psalm-baseline.xml",
+		"psalm": "psalm --no-cache",
+		"psalm:update-baseline": "psalm --threads=1 --update-baseline",
+		"psalm:update-baseline:force": "psalm --threads=1 --update-baseline --set-baseline=tests/psalm-baseline.xml",
 		"test:unit": "phpunit --config tests/phpunit.xml",
 		"openapi": "generate-spec --verbose --allow-missing-docs --continue-on-error"
 	},
 	"require": {
-		"php": "^8.0 | ^8.1 | ^8.2 | ^8.3 | ^8.4",
-		"ext-xml": "*"
+		"php": "^8.1 | ^8.2 | ^8.3 | ^8.4",
+		"ext-xml": "*",
+		"bamarni/composer-bin-plugin": "^1.8"
 	},
 	"require-dev": {
 		"nextcloud/coding-standard": "^1.1",
-		"psalm/phar": "^6",
 		"nextcloud/ocp": "dev-master",
-		"phpunit/phpunit": "^9.5",
+		"phpunit/phpunit": "^12",
 		"nextcloud/openapi-extractor": "^1.7.0"
+	},
+	"config": {
+		"optimize-autoloader": true,
+		"classmap-authoritative": false,
+		"allow-plugins": {
+			"bamarni/composer-bin-plugin": true
+		}
+	},
+	"extra": {
+		"bamarni-bin": {
+			"bin-links": true,
+			"target-directory": "vendor-bin",
+			"forward-command": true
+		}
 	}
 }
