@@ -24,74 +24,57 @@
 
 namespace OCA\PhoneTrack\Activity;
 
+use OCA\PhoneTrack\AppInfo\Application;
+use OCP\Activity\ActivitySettings;
 use OCP\IL10N;
 
-class Setting implements \OCP\Activity\ISetting {
+class Setting extends ActivitySettings {
 
-	/** @var IL10N */
-	protected $l;
-
-	/**
-	 * @param IL10N $l
-	 */
-	public function __construct(IL10N $l) {
-		$this->l = $l;
+	public function __construct(
+		protected IL10N $l,
+	) {
 	}
 
-	/**
-	 * @return string Lowercase a-z and underscore only identifier
-	 * @since 11.0.0
-	 */
 	public function getIdentifier() {
 		return 'phonetrack';
 	}
 
-	/**
-	 * @return string A translated string
-	 * @since 11.0.0
-	 */
 	public function getName() {
-		return $this->l->t('Changes in the <strong>PhoneTrack app</strong>');
+		return $this->l->t('A <strong>PhoneTrack session</strong> has been shared or unshared');
 	}
 
 	/**
-	 * @return int whether the filter should be rather on the top or bottom of
-	 *             the admin section. The filters are arranged in ascending order of the
-	 *             priority values. It is required to return a value between 0 and 100.
-	 * @since 11.0.0
+	 * {@inheritdoc}
 	 */
+	#[\Override]
+	public function getGroupIdentifier(): string {
+		return Application::APP_ID;
+	}
+
+	/**
+	 * {@inheritdoc}
+	 */
+	#[\Override]
+	public function getGroupName(): string {
+		return $this->l->t('Phonetrack');
+	}
+
 	public function getPriority() {
 		return 95;
 	}
 
-	/**
-	 * @return bool True when the option can be changed for the stream
-	 * @since 11.0.0
-	 */
 	public function canChangeStream() {
 		return true;
 	}
 
-	/**
-	 * @return bool True when the option can be changed for the stream
-	 * @since 11.0.0
-	 */
 	public function isDefaultEnabledStream() {
 		return true;
 	}
 
-	/**
-	 * @return bool True when the option can be changed for the mail
-	 * @since 11.0.0
-	 */
 	public function canChangeMail() {
 		return true;
 	}
 
-	/**
-	 * @return bool True when the option can be changed for the stream
-	 * @since 11.0.0
-	 */
 	public function isDefaultEnabledMail() {
 		return false;
 	}
