@@ -10,41 +10,65 @@
  * @copyright Julien Veyssier 2017
  */
 
-/**
- * Create your routes in here. The name is the lowercase name of the controller
- * without the controller part, the stuff after the hash is the method.
- * e.g. page#index -> OCA\PhoneTrack\Controller\PageController->index()
- *
- * The controller class has to be registered in the application.php file since
- * it's instantiated in there
- */
 return [
 	'routes' => [
-		['name' => 'page#index', 'url' => '/', 'verb' => 'GET'],
-		['name' => 'page#publicWebLog', 'url' => '/publicWebLog/{token}/{devicename}', 'verb' => 'GET'],
-		['name' => 'page#publicSessionWatch', 'url' => '/publicSessionWatch/{publicviewtoken}', 'verb' => 'GET'],
-		['name' => 'page#createSession', 'url' => '/createSession', 'verb' => 'POST'],
-		['name' => 'page#deleteSession', 'url' => '/deleteSession', 'verb' => 'POST'],
-		['name' => 'page#deletePoints', 'url' => '/deletePoints', 'verb' => 'POST'],
-		['name' => 'page#updatePoint', 'url' => '/updatePoint', 'verb' => 'POST'],
+		['name' => 'page#index', 'url' => '/vue', 'verb' => 'GET'],
+		['name' => 'page#getSessions', 'url' => '/sessions', 'verb' => 'GET'],
+		['name' => 'page#createSession', 'url' => '/session', 'verb' => 'POST'],
+		['name' => 'page#importSession', 'url' => '/import-session', 'verb' => 'POST'],
+		['name' => 'page#deleteSession', 'url' => '/session/{sessionId}', 'verb' => 'DELETE'],
+		['name' => 'page#updateSession', 'url' => '/session/{sessionId}', 'verb' => 'PUT'],
+		['name' => 'page#downloadSession', 'url' => '/session/{sessionId}/download', 'verb' => 'GET'],
+		['name' => 'page#exportDevice', 'url' => '/session/{sessionId}/device/{deviceId}/export', 'verb' => 'POST'],
+		['name' => 'page#deleteDevice', 'url' => '/session/{sessionId}/device/{deviceId}', 'verb' => 'DELETE'],
+		['name' => 'page#updateDevice', 'url' => '/session/{sessionId}/device/{deviceId}', 'verb' => 'PUT'],
+		['name' => 'page#addDeviceReservation', 'url' => '/session/{sessionId}/device-name/{deviceName}', 'verb' => 'POST'],
+		['name' => 'page#createPublicShare', 'url' => '/session/{sessionId}/pub-share', 'verb' => 'POST'],
+		['name' => 'page#updatePublicShare', 'url' => '/session/{sessionId}/pub-share/{pubShareId}', 'verb' => 'PUT'],
+		['name' => 'page#deletePublicShare', 'url' => '/session/{sessionId}/pub-share/{pubShareId}', 'verb' => 'DELETE'],
+		['name' => 'page#createShare', 'url' => '/session/{sessionId}/share', 'verb' => 'POST'],
+		['name' => 'page#deleteShare', 'url' => '/session/{sessionId}/share/{shareId}', 'verb' => 'DELETE'],
+		['name' => 'page#createGeofence', 'url' => '/session/{sessionId}/device/{deviceId}/geofence', 'verb' => 'POST'],
+		['name' => 'page#updateGeofence', 'url' => '/session/{sessionId}/device/{deviceId}/geofence/{geofenceId}', 'verb' => 'PUT'],
+		['name' => 'page#deleteGeofence', 'url' => '/session/{sessionId}/device/{deviceId}/geofence/{geofenceId}', 'verb' => 'DELETE'],
+		['name' => 'page#createProxim', 'url' => '/session/{sessionId1}/device/{deviceId1}/proxim', 'verb' => 'POST'],
+		['name' => 'page#updateProxim', 'url' => '/session/{sessionId1}/device/{deviceId1}/proxim/{proximId}', 'verb' => 'PUT'],
+		['name' => 'page#deleteProxim', 'url' => '/session/{sessionId1}/device/{deviceId1}/proxim/{proximId}', 'verb' => 'DELETE'],
+		['name' => 'page#getDevicePoints', 'url' => '/session/{sessionId}/device/{deviceId}/points', 'verb' => 'GET'],
+		['name' => 'page#deletePoint', 'url' => '/session/{sessionId}/device/{deviceId}/point/{pointId}', 'verb' => 'DELETE'],
+		['name' => 'page#updatePoint', 'url' => '/session/{sessionId}/device/{deviceId}/point/{pointId}', 'verb' => 'PUT'],
+		['name' => 'page#getSvgFromApp', 'url' => '/svg/{fileName}', 'verb' => 'GET'],
+
+		['name' => 'publicShare#index', 'url' => '/s/{token}', 'verb' => 'GET'],
+		['name' => 'publicShare#getDevicePoints', 'url' => '/s/{shareToken}/device/{deviceId}/points', 'verb' => 'GET'],
+
+		['name' => 'oldPage#index', 'url' => '/', 'verb' => 'GET'],
+		['name' => 'oldPage#publicWebLog', 'url' => '/publicWebLog/{token}/{devicename}', 'verb' => 'GET'],
+		['name' => 'oldPage#publicSessionWatch', 'url' => '/publicSessionWatch/{publicviewtoken}', 'verb' => 'GET'],
+		['name' => 'oldPage#createSession', 'url' => '/createSession', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteSession', 'url' => '/deleteSession', 'verb' => 'POST'],
+		['name' => 'oldPage#deletePoints', 'url' => '/deletePoints', 'verb' => 'POST'],
+		['name' => 'oldPage#updatePoint', 'url' => '/updatePoint', 'verb' => 'POST'],
+		['name' => 'oldPage#getSessions', 'url' => '/getSessions', 'verb' => 'POST'],
+		['name' => 'oldPage#APIgetSessions', 'url' => '/api/getsessions', 'verb' => 'GET'],
+		['name' => 'oldPage#APIgetLastPositionsPublic', 'url' => '/api/getlastpositions/{sessionId}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIgetLastDevicePositionPublic', 'url' => '/api/getlastpositions/{sessionId}/{deviceName}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIgetPositionsPublic', 'url' => '/api/getpositions/{sessionid}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIgetLastPositionsUser', 'url' => '/api/getuserlastpositions/{sessionid}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIgetPositionsUser', 'url' => '/api/getuserpositions/{sessionid}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIshareDevice', 'url' => '/api/sharedevice/{sessionid}/{devicename}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIcreateSession', 'url' => '/api/createsession/{sessionname}', 'verb' => 'GET'],
+		['name' => 'oldPage#APIPing', 'url' => '/api/ping', 'verb' => 'GET'],
+		['name' => 'oldPage#importSession', 'url' => '/importSession', 'verb' => 'POST'],
+		['name' => 'oldPage#renameSession', 'url' => '/renameSession', 'verb' => 'POST'],
+		['name' => 'oldPage#renameDevice', 'url' => '/renameDevice', 'verb' => 'POST'],
+		['name' => 'oldPage#setDeviceAlias', 'url' => '/setDeviceAlias', 'verb' => 'POST'],
+		['name' => 'oldPage#reaffectDevice', 'url' => '/reaffectDevice', 'verb' => 'POST'],
+		['name' => 'oldPage#setDeviceColor', 'url' => '/setDeviceColor', 'verb' => 'POST'],
+		['name' => 'oldPage#setDeviceShape', 'url' => '/setDeviceShape', 'verb' => 'POST'],
+
 		['name' => 'log#addPoint', 'url' => '/addPoint', 'verb' => 'POST'],
-		['name' => 'page#getSessions', 'url' => '/getSessions', 'verb' => 'POST'],
-		['name' => 'page#APIgetSessions', 'url' => '/api/getsessions', 'verb' => 'GET'],
-		['name' => 'page#APIgetLastPositionsPublic', 'url' => '/api/getlastpositions/{sessionId}', 'verb' => 'GET'],
-		['name' => 'page#APIgetLastDevicePositionPublic', 'url' => '/api/getlastpositions/{sessionId}/{deviceName}', 'verb' => 'GET'],
-		['name' => 'page#APIgetPositionsPublic', 'url' => '/api/getpositions/{sessionid}', 'verb' => 'GET'],
-		['name' => 'page#APIgetLastPositionsUser', 'url' => '/api/getuserlastpositions/{sessionid}', 'verb' => 'GET'],
-		['name' => 'page#APIgetPositionsUser', 'url' => '/api/getuserpositions/{sessionid}', 'verb' => 'GET'],
-		['name' => 'page#APIshareDevice', 'url' => '/api/sharedevice/{sessionid}/{devicename}', 'verb' => 'GET'],
-		['name' => 'page#APIcreateSession', 'url' => '/api/createsession/{sessionname}', 'verb' => 'GET'],
-		['name' => 'page#APIPing', 'url' => '/api/ping', 'verb' => 'GET'],
-		['name' => 'page#importSession', 'url' => '/importSession', 'verb' => 'POST'],
-		['name' => 'page#renameSession', 'url' => '/renameSession', 'verb' => 'POST'],
-		['name' => 'page#renameDevice', 'url' => '/renameDevice', 'verb' => 'POST'],
-		['name' => 'page#setDeviceAlias', 'url' => '/setDeviceAlias', 'verb' => 'POST'],
-		['name' => 'page#reaffectDevice', 'url' => '/reaffectDevice', 'verb' => 'POST'],
-		['name' => 'page#setDeviceColor', 'url' => '/setDeviceColor', 'verb' => 'POST'],
-		['name' => 'page#setDeviceShape', 'url' => '/setDeviceShape', 'verb' => 'POST'],
+		['name' => 'log#addPoint2', 'url' => '/session/{sessionId}/device/{deviceId}/point', 'verb' => 'POST'],
 		['name' => 'log#logGet', 'url' => '/logGet/{token}/{devicename}', 'verb' => 'GET'],
 		['name' => 'log#logPost', 'url' => '/logPost/{token}/{devicename}', 'verb' => 'POST'],
 		['name' => 'log#logPostMultiple', 'url' => '/logPostMultiple/{token}/{devicename}', 'verb' => 'POST'],
@@ -59,35 +83,56 @@ return [
 		['name' => 'log#logTraccar', 'url' => '/log/traccar/{token}/{deviceName}', 'verb' => 'POST', 'defaults' => ['deviceName' => '']],
 		['name' => 'log#logOpengts', 'url' => '/log/opengts/{token}/{devicename}', 'verb' => 'GET', 'defaults' => ['devicename' => '']],
 		['name' => 'log#logOpengtsPost', 'url' => '/log/opengts/{token}/{devicename}', 'verb' => 'POST'],
-		['name' => 'page#track', 'url' => '/track', 'verb' => 'POST'],
-		['name' => 'page#publicWebLogTrack', 'url' => '/publicWebLogTrack', 'verb' => 'POST'],
-		['name' => 'page#publicViewTrack', 'url' => '/publicViewTrack', 'verb' => 'POST'],
-		['name' => 'page#setSessionPublic', 'url' => '/setSessionPublic', 'verb' => 'POST'],
-		['name' => 'page#setSessionLocked', 'url' => '/setSessionLocked', 'verb' => 'POST'],
-		['name' => 'page#setSessionAutoExport', 'url' => '/setSessionAutoExport', 'verb' => 'POST'],
-		['name' => 'page#setSessionAutoPurge', 'url' => '/setSessionAutoPurge', 'verb' => 'POST'],
-		['name' => 'page#addUserShare', 'url' => '/addUserShare', 'verb' => 'POST'],
-		['name' => 'page#deleteUserShare', 'url' => '/deleteUserShare', 'verb' => 'POST'],
-		['name' => 'page#addPublicShare', 'url' => '/addPublicShare', 'verb' => 'POST'],
-		['name' => 'page#deletePublicShare', 'url' => '/deletePublicShare', 'verb' => 'POST'],
-		['name' => 'page#setPublicShareDevice', 'url' => '/setPublicShareDevice', 'verb' => 'POST'],
-		['name' => 'page#setPublicShareLastOnly', 'url' => '/setPublicShareLastOnly', 'verb' => 'POST'],
-		['name' => 'page#setPublicShareGeofencify', 'url' => '/setPublicShareGeofencify', 'verb' => 'POST'],
-		['name' => 'page#addNameReservation', 'url' => '/addNameReservation', 'verb' => 'POST'],
-		['name' => 'page#deleteNameReservation', 'url' => '/deleteNameReservation', 'verb' => 'POST'],
-		['name' => 'page#addFiltersBookmark', 'url' => '/addFiltersBookmark', 'verb' => 'POST'],
-		['name' => 'page#deleteFiltersBookmark', 'url' => '/deleteFiltersBookmark', 'verb' => 'POST'],
-		['name' => 'page#addGeofence', 'url' => '/addGeofence', 'verb' => 'POST'],
-		['name' => 'page#deleteGeofence', 'url' => '/deleteGeofence', 'verb' => 'POST'],
-		['name' => 'page#addProxim', 'url' => '/addProxim', 'verb' => 'POST'],
-		['name' => 'page#deleteProxim', 'url' => '/deleteProxim', 'verb' => 'POST'],
-		['name' => 'page#export', 'url' => '/export', 'verb' => 'POST'],
-		['name' => 'page#deleteDevice', 'url' => '/deleteDevice', 'verb' => 'POST'],
-		['name' => 'page#getUserList', 'url' => '/getUserList', 'verb' => 'POST'],
-		['name' => 'utils#setPointQuota', 'url' => '/setPointQuota', 'verb' => 'POST'],
-		['name' => 'utils#addTileServer', 'url' => '/addTileServer', 'verb' => 'POST'],
-		['name' => 'utils#deleteTileServer', 'url' => '/deleteTileServer', 'verb' => 'POST'],
+
+		['name' => 'oldPage#track', 'url' => '/track', 'verb' => 'POST'],
+		['name' => 'oldPage#publicWebLogTrack', 'url' => '/publicWebLogTrack', 'verb' => 'POST'],
+		['name' => 'oldPage#publicViewTrack', 'url' => '/publicViewTrack', 'verb' => 'POST'],
+		['name' => 'oldPage#setSessionPublic', 'url' => '/setSessionPublic', 'verb' => 'POST'],
+		['name' => 'oldPage#setSessionLocked', 'url' => '/setSessionLocked', 'verb' => 'POST'],
+		['name' => 'oldPage#setSessionAutoExport', 'url' => '/setSessionAutoExport', 'verb' => 'POST'],
+		['name' => 'oldPage#setSessionAutoPurge', 'url' => '/setSessionAutoPurge', 'verb' => 'POST'],
+		['name' => 'oldPage#addUserShare', 'url' => '/addUserShare', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteUserShare', 'url' => '/deleteUserShare', 'verb' => 'POST'],
+		['name' => 'oldPage#addPublicShare', 'url' => '/addPublicShare', 'verb' => 'POST'],
+		['name' => 'oldPage#deletePublicShare', 'url' => '/deletePublicShare', 'verb' => 'POST'],
+		['name' => 'oldPage#setPublicShareDevice', 'url' => '/setPublicShareDevice', 'verb' => 'POST'],
+		['name' => 'oldPage#setPublicShareLastOnly', 'url' => '/setPublicShareLastOnly', 'verb' => 'POST'],
+		['name' => 'oldPage#setPublicShareGeofencify', 'url' => '/setPublicShareGeofencify', 'verb' => 'POST'],
+		['name' => 'oldPage#addNameReservation', 'url' => '/addNameReservation', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteNameReservation', 'url' => '/deleteNameReservation', 'verb' => 'POST'],
+		['name' => 'oldPage#addFiltersBookmark', 'url' => '/addFiltersBookmark', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteFiltersBookmark', 'url' => '/deleteFiltersBookmark', 'verb' => 'POST'],
+		['name' => 'oldPage#addGeofence', 'url' => '/addGeofence', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteGeofence', 'url' => '/deleteGeofence', 'verb' => 'POST'],
+		['name' => 'oldPage#addProxim', 'url' => '/addProxim', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteProxim', 'url' => '/deleteProxim', 'verb' => 'POST'],
+		['name' => 'oldPage#export', 'url' => '/export', 'verb' => 'POST'],
+		['name' => 'oldPage#deleteDevice', 'url' => '/deleteDevice', 'verb' => 'POST'],
+		['name' => 'oldPage#getUserList', 'url' => '/getUserList', 'verb' => 'POST'],
+
+		['name' => 'utils#oldAddTileServer', 'url' => '/old/addTileServer', 'verb' => 'POST'],
+		['name' => 'utils#oldDeleteTileServer', 'url' => '/old/deleteTileServer', 'verb' => 'POST'],
+
+		['name' => 'utils#addTileServer', 'url' => '/tileservers', 'verb' => 'POST'],
+		['name' => 'utils#updateTileServer', 'url' => '/tileservers/{id}', 'verb' => 'PUT'],
+		['name' => 'utils#deleteTileServer', 'url' => '/tileservers/{id}', 'verb' => 'DELETE'],
+		['name' => 'utils#adminAddTileServer', 'url' => '/admin/tileservers', 'verb' => 'POST'],
+		['name' => 'utils#AdminUpdateTileServer', 'url' => '/admin/tileservers/{id}', 'verb' => 'PUT'],
+		['name' => 'utils#AdminDeleteTileServer', 'url' => '/admin/tileservers/{id}', 'verb' => 'DELETE'],
 		['name' => 'utils#getOptionsValues', 'url' => '/getOptionsValues', 'verb' => 'POST'],
 		['name' => 'utils#saveOptionValue', 'url' => '/saveOptionValue', 'verb' => 'POST'],
+		['name' => 'utils#saveOptionValues', 'url' => '/saveOptionValues', 'verb' => 'PUT'],
+		['name' => 'utils#setAdminConfig', 'url' => '/admin-config', 'verb' => 'PUT'],
+		['name' => 'utils#setSensitiveAdminConfig', 'url' => '/admin-config/sensitive', 'verb' => 'PUT'],
+
+		['name' => 'map#getRasterTile', 'url' => '/tiles/{service}/{x}/{y}/{z}', 'verb' => 'GET'],
+		['name' => 'map#nominatimSearch', 'url' => '/nominatim/search', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerStyle', 'url' => '/maptiler/maps/{version}/style.json', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerFont', 'url' => '/maptiler/fonts/{fontstack}/{range}.pbf', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerTiles', 'url' => '/maptiler/tiles/{version}/tiles.json', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerTile', 'url' => '/maptiler/tiles/{version}/{z}/{x}/{y}.{ext}', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerSpriteNoSize', 'url' => '/maptiler/maps/{version}/sprite.{ext}', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerSprite', 'url' => '/maptiler/maps/{version}/sprite{size}.{ext}', 'verb' => 'GET'],
+		['name' => 'map#getMapTilerResource', 'url' => '/maptiler/resources/{name}', 'verb' => 'GET'],
 	]
 ];
