@@ -1,17 +1,9 @@
 <?php
 
-/**
- * Nextcloud - phonetrack
- *
- * This file is licensed under the Affero General Public License version 3 or
- * later. See the COPYING file.
- *
- * @author Julien Veyssier <eneiluj@posteo.net
- * @copyright Julien Veyssier 2019
- */
-
 namespace OCA\PhoneTrack\Db;
 
+use OCP\AppFramework\Db\DoesNotExistException;
+use OCP\AppFramework\Db\MultipleObjectsReturnedException;
 use OCP\AppFramework\Db\QBMapper;
 use OCP\DB\Exception;
 use OCP\DB\QueryBuilder\IQueryBuilder;
@@ -26,7 +18,7 @@ class DeviceMapper extends QBMapper {
 		parent::__construct($db, 'phonetrack_devices', Device::class);
 	}
 
-	public function find($id) {
+	public function find(int $id): Device {
 		$qb = $this->db->getQueryBuilder();
 
 		$qb->select('*')
@@ -38,19 +30,139 @@ class DeviceMapper extends QBMapper {
 		return $this->findEntity($qb);
 	}
 
-	public function findBySessionId(string $sessionId) {
+	/**
+	 * @param int $sessionId
+	 * @return Device[]
+	 * @throws Exception
+	 */
+	public function findBySessionId(int $sessionId): array {
 		$qb = $this->db->getQueryBuilder();
 
 		$qb->select('*')
 			->from($this->getTableName())
 			->where(
-				$qb->expr()->eq('sessionid', $qb->createNamedParameter($sessionId, IQueryBuilder::PARAM_STR))
+				$qb->expr()->eq('session_id', $qb->createNamedParameter($sessionId, IQueryBuilder::PARAM_INT))
 			);
 
 		return $this->findEntities($qb);
 	}
 
-	public function deletePointsOlderThan(int $deviceId, int $timestamp) {
+	/**
+	 * @param string $sessionToken
+	 * @param int $deviceId
+	 * @return Device
+	 * @throws DoesNotExistException
+	 * @throws Exception
+	 * @throws MultipleObjectsReturnedException
+	 */
+	public function getBySessionTokenAndDeviceId(string $sessionToken, int $deviceId): Device {
+		$qb = $this->db->getQueryBuilder();
+
+		$qb->select('*')
+			->from($this->getTableName())
+			->where(
+				$qb->expr()->eq('session_token', $qb->createNamedParameter($sessionToken, IQueryBuilder::PARAM_STR))
+			)
+			->andWhere(
+				$qb->expr()->eq('id', $qb->createNamedParameter($deviceId, IQueryBuilder::PARAM_INT))
+			);
+
+		return $this->findEntity($qb);
+	}
+
+	/**
+	 * @param int $sessionId
+	 * @param int $deviceId
+	 * @return Device
+	 * @throws DoesNotExistException
+	 * @throws Exception
+	 * @throws MultipleObjectsReturnedException
+	 */
+	public function getBySessionIdAndDeviceId(int $sessionId, int $deviceId): Device {
+		$qb = $this->db->getQueryBuilder();
+
+		$qb->select('*')
+			->from($this->getTableName())
+			->where(
+				$qb->expr()->eq('session_id', $qb->createNamedParameter($sessionId, IQueryBuilder::PARAM_INT))
+			)
+			->andWhere(
+				$qb->expr()->eq('id', $qb->createNamedParameter($deviceId, IQueryBuilder::PARAM_INT))
+			);
+
+		return $this->findEntity($qb);
+	}
+
+	/**
+	 * @param string $sessionToken
+	 * @param string $name
+	 * @return Device
+	 * @throws Exception
+	 * @throws DoesNotExistException
+	 * @throws MultipleObjectsReturnedException
+	 */
+	public function getByName(string $sessionToken, string $name): Device {
+		$qb = $this->db->getQueryBuilder();
+
+		$qb->select('*')
+			->from($this->getTableName())
+			->where(
+				$qb->expr()->eq('session_token', $qb->createNamedParameter($sessionToken, IQueryBuilder::PARAM_STR))
+			)
+			->andWhere(
+				$qb->expr()->eq('name', $qb->createNamedParameter($name, IQueryBuilder::PARAM_STR))
+			);
+
+		return $this->findEntity($qb);
+	}
+
+	/**
+	 * @param string $sessionToken
+	 * @param string $nameToken
+	 * @return Device
+	 * @throws DoesNotExistException
+	 * @throws Exception
+	 * @throws MultipleObjectsReturnedException
+	 */
+	public function getByNameToken(string $sessionToken, string $nameToken): Device {
+		$qb = $this->db->getQueryBuilder();
+
+		$qb->select('*')
+			->from($this->getTableName())
+			->where(
+				$qb->expr()->eq('session_token', $qb->createNamedParameter($sessionToken, IQueryBuilder::PARAM_STR))
+			)
+			->andWhere(
+				$qb->expr()->eq('nametoken', $qb->createNamedParameter($nameToken, IQueryBuilder::PARAM_STR))
+			);
+
+		return $this->findEntity($qb);
+	}
+
+	public function deleteDevice(string $sessionToken, int $deviceId): void {
+		$this->deleteDevicePoints($deviceId);
+
+		$qb = $this->db->getQueryBuilder();
+		$qb->delete('phonetrack_devices')
+			->where(
+				$qb->expr()->eq('session_token', $qb->createNamedParameter($sessionToken, IQueryBuilder::PARAM_STR))
+			)
+			->andWhere(
+				$qb->expr()->eq('id', $qb->createNamedParameter($deviceId, IQueryBuilder::PARAM_INT))
+			);
+		$qb->executeStatement();
+	}
+
+	public function deleteDevicePoints(int $deviceId): int {
+		$qb = $this->db->getQueryBuilder();
+		$qb->delete('phonetrack_points')
+			->where(
+				$qb->expr()->eq('deviceid', $qb->createNamedParameter($deviceId, IQueryBuilder::PARAM_INT))
+			);
+		return $qb->executeStatement();
+	}
+
+	public function deletePointsOlderThan(int $deviceId, int $timestamp): int {
 		$qb = $this->db->getQueryBuilder();
 		$qb->delete('phonetrack_points')
 			->where(
@@ -59,16 +171,16 @@ class DeviceMapper extends QBMapper {
 			->andWhere(
 				$qb->expr()->lt('timestamp', $qb->createNamedParameter($timestamp, IQueryBuilder::PARAM_INT))
 			);
-		$qb->executeStatement();
+		return $qb->executeStatement();
 	}
 
-	public function countDevicesPerSession(string $token): int {
+	public function countDevicesPerSession(int $sessionId): int {
 		$qb = $this->db->getQueryBuilder();
 
 		$qb->selectAlias($qb->createFunction('COUNT(*)'), 'count_devs')
 			->from($this->getTableName())
 			->where(
-				$qb->expr()->eq('sessionid', $qb->createNamedParameter($token))
+				$qb->expr()->eq('session_id', $qb->createNamedParameter($sessionId, IQueryBuilder::PARAM_INT))
 			);
 
 		$req = $qb->executeQuery();
@@ -129,9 +241,7 @@ class DeviceMapper extends QBMapper {
 		}
 
 		$req = $qb->executeQuery();
-		$points = $req->fetchAll();
-		$qb->resetQueryParts();
-		return $points;
+		return $req->fetchAll();
 	}
 
 	/**
@@ -167,5 +277,45 @@ class DeviceMapper extends QBMapper {
 			}
 		}
 		return $qb;
+	}
+
+	public function getSessionOwnerOfDevice(int $deviceId): string {
+		/*
+			SELECT ' . $this->dbdblquotes . 'user' . $this->dbdblquotes . '
+			FROM *PREFIX*phonetrack_devices
+			INNER JOIN *PREFIX*phonetrack_sessions
+				ON *PREFIX*phonetrack_devices.session_token=*PREFIX*phonetrack_sessions.token
+			WHERE *PREFIX*phonetrack_devices.id=' . $this->db_quote_escape_string($deviceId) . ' ;';
+		 */
+		$qb = $this->db->getQueryBuilder();
+		$qb->select('user')
+			->from($this->getTableName(), 'device')
+			->innerJoin('device', 'phonetrack_sessions', 'session', $qb->expr()->eq('device.session_id', 'session.id'))
+			->where(
+				$qb->expr()->eq('device.id', $qb->createNamedParameter($deviceId, IQueryBuilder::PARAM_INT))
+			);
+
+		$res = $qb->executeQuery();
+		return (string)$res->fetchOne();
+	}
+
+	/**
+	 * @param int $sessionId
+	 * @return Device[]
+	 * @throws Exception
+	 */
+	public function getDevicesWithPointsInSession(int $sessionId): array {
+		$qb = $this->db->getQueryBuilder();
+		$qb->select('dev.*')
+			->from($this->getTableName(), 'dev')
+			->from('phonetrack_points', 'po')
+			->where(
+				$qb->expr()->eq('dev.session_id', $qb->createNamedParameter($sessionId, IQueryBuilder::PARAM_INT))
+			)
+			->andWhere(
+				$qb->expr()->eq('dev.id', 'po.deviceid')
+			)
+			->groupBy('dev.id');
+		return $this->findEntities($qb);
 	}
 }
