@@ -4,24 +4,33 @@
  */
 import { createAppConfig } from '@nextcloud/vite-config'
 import eslint from 'vite-plugin-eslint'
-import stylelint from 'vite-plugin-stylelint'
+// import stylelint from 'vite-plugin-stylelint'
 
 const isProduction = process.env.NODE_ENV === 'production'
 
 export default createAppConfig({
-	admin: 'src/admin.js',
+	adminSettings: 'src/adminSettings.js',
 	phonetrack: 'src/phonetrack.js',
+    mainVue: 'src/mainVue.js',
 }, {
 	config: {
 		css: {
 			modules: {
 				localsConvention: 'camelCase',
 			},
+            preprocessorOptions: {
+				scss: {
+					api: 'modern-compiler',
+				},
+			},
 		},
 		plugins: [
 			eslint(),
-			stylelint(),
+			// stylelint(),
 		],
+        build: {
+			cssCodeSplit: true,
+		},
 	},
 	inlineCSS: { relativeCSSInjection: true },
 	minify: isProduction,
