@@ -1,8 +1,8 @@
 const webpack = require('webpack')
 const path = require('path')
 const webpackConfig = require('@nextcloud/webpack-vue-config')
+// const StyleLintPlugin = require('stylelint-webpack-plugin')
 const ESLintPlugin = require('eslint-webpack-plugin')
-const StyleLintPlugin = require('stylelint-webpack-plugin')
 
 const buildMode = process.env.NODE_ENV
 const isDev = buildMode === 'development'
@@ -15,8 +15,9 @@ webpackConfig.stats = {
 
 const appId = 'phonetrack'
 webpackConfig.entry = {
-    admin: { import: path.join(__dirname, 'src', 'admin.js'), filename: appId + '-admin.js' },
+    adminSettings: { import: path.join(__dirname, 'src', 'adminSettings.js'), filename: appId + '-adminSettings.js' },
     phonetrack: { import: path.join(__dirname, 'src', 'phonetrack.js'), filename: appId + '-phonetrack.js' },
+	mainVue: { import: path.join(__dirname, 'src', 'mainVue.js'), filename: appId + '-mainVue.js' },
 }
 
 webpackConfig.plugins.push(
@@ -28,19 +29,19 @@ webpackConfig.plugins.push(
         'window.$': 'jquery'
     })
 )
-
 webpackConfig.plugins.push(
-	new ESLintPlugin({
-		extensions: ['js', 'vue'],
-		files: 'src',
-		failOnError: !isDev,
-	})
+    new ESLintPlugin({
+        extensions: ['js', 'vue'],
+        files: 'src',
+        failOnError: !isDev,
+		configType: 'eslintrc',
+    })
 )
-webpackConfig.plugins.push(
-	new StyleLintPlugin({
-		files: 'src/**/*.{css,scss,vue}',
-		failOnError: !isDev,
-	}),
-)
+//webpackConfig.plugins.push(
+//	new StyleLintPlugin({
+//		files: 'src/**/*.{css,scss,vue}',
+//		failOnError: !isDev,
+//	}),
+//)
 
 module.exports = webpackConfig
